\name{getL2normL2deriv}
\alias{getL2normL2deriv}

\title{Calculation of L2 norm of L2derivative}
\description{
  Function to calculate the L2 norm of the L2derivative in a smooth parametric model.
}
\usage{getL2normL2deriv(aFinfo, cent, ...)}
%\details{}
\arguments{
  \item{aFinfo}{trace of the Fisher information}
  \item{cent}{centering}
  \item{...}{further arguments (not used at the moment)}
}

\value{L2 norm of the L2derivative}
\author{Peter Ruckdeschel \email{peter.ruckdeschel@uni-oldenburg.de}}
%\note{}
\examples{
##
}
\concept{L2norm}
\keyword{robust}
